Move StickYaw class into a globally available library
