+++ src/lib/stick_yaw/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2018 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2019 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -31,9 +31,7 @@
 #
 ############################################################################
 
-px4_add_library(FlightTaskManualAltitude
-	FlightTaskManualAltitude.cpp
+px4_add_library(StickYaw
+	StickYaw.cpp
 )
-
-target_link_libraries(FlightTaskManualAltitude PUBLIC FlightTask FlightTaskUtility StickYaw)
-target_include_directories(FlightTaskManualAltitude PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+target_include_directories(StickYaw PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
